🧹 Janitor: rename sourceprovider common package to sourcecache
AGENTS.md forbids utils/common package names. The package only provides
EnsureCachedDir for module source caching — rename to sourcecache and
update the github provider import. Behavior unchanged.
