+++ internal/modfile/sourceprovider/sourcecache/cache.go
@@ -1,4 +1,4 @@
-package common
+package sourcecache
 
 import (
 	"context"
